requirements.txt
streamlit>=1.18
openai>=1.0.0
gtts>=2.3.0
python-multipart>=0.0.5
groq>=0.5.0
python-dotenv>=1.0.0